refactor(pages): Rename victim memory page to simply page #9
