+++ src/memories/pages/index.js
@@ -1,2 +1,2 @@
+export { default as Memory } from './memory'
 export { default as MemoryCreate } from './memory-create'
-export { default as MemoryVictim } from './memory-victim'
