+++ src/memories/pages/memory/page.connected.js
